dagster_app/requirements.txt
dagster==1.9.10
dagster-webserver==1.9.10
dagster-postgres==0.25.10
httpx>=0.27
psycopg2-binary>=2.9
